# math/big: simplify shlVU_g and shrVU_g